move privateca ref to a separate file
